telegram/tg-pusher/requirements.txt
aiogram>=3.4
aiosqlite
python-dotenv
croniter
